FASTQ output

+++ src/c++/lib/genome/QualityModel.cpp
@@ -874,7 +874,7 @@ void ErrorModel::getQualityAndRandomError( boost::mt19937& randomGen, const unsi
 
     // Apply quality drop due to phasing, using an additive strategy
     // This quality drop was calculated as part of the previous "applyQualityDrop" methods
-    if (quality > clusterErrorModelContext.phasingContext.qualityDrop)
+    if ((int)quality > clusterErrorModelContext.phasingContext.qualityDrop)
     {
       quality -= clusterErrorModelContext.phasingContext.qualityDrop;
     }
